Separated out Comments functionality into different hooks file and Comments list component
